update lb to match with fw
now the load_balancer will assign a server to response either the request dst is the load_balancer or one of the servers.
so now there is no need to configure firewall to allow traffic to load_balancer and can use h1 curl 100.0.0.40:80. load_balancer will still randomly assign a server to response.
curl 100.0.0.30 still works but h1 and h2 cant get response due to the firwall.